thrift: 0.18.1 -> 0.21.0
I'm fetching from Github instead of the apache server since the release
from apache.org appears to be missing the file
"lib/cpp/test/Thrift5272.thrift". As such, it will not build. For
reasons that are unclear, the mirror on github appears to have that file
and will build.

I remove all the patches since it appears to build on aarch64-darwin
without the patches.

I removed the disabled tests since it appears to build on aarch64-darwin
(sandbox = true) without those tests being disabled.