cmake: drop code referencing `/var/empty` from the purity patch
These are almost all dead code – some references to `/bin` and
`/sbin` and some Windows paths notwithstanding – and dropping them
entirely makes the patch shorter and more effective.

It is likely that a lot of this patch is still superfluous; we
don’t really care about the purity of unsandboxed builds on Linux,
for instance, so any code path that is never taken on Darwin could
probably be dropped. But this provides a conservative base to drop the
problematic hook without compromise for CMake‐distributed modules. Of
course, the long‐term solution should be proper sandboxing on macOS.